refactor: CopilotChat.WebApi.Auth.Specialization => CopilotChat.WebApi.Auth.Specializations
