+++ webapi/Auth/Specializations/LinkedToChatAuthorizationHandler.cs
@@ -5,7 +5,7 @@ using CopilotChat.WebApi.Context;
 using CopilotChat.WebApi.Storage;
 using Microsoft.AspNetCore.Authorization;
 
-namespace CopilotChat.WebApi.Auth.Specialization;
+namespace CopilotChat.WebApi.Auth.Specializations;
 
 /// <summary>
 /// Class implementing "authorization" that validates the user has access to a specialization linked to the chat being targeted.
