+++ webapi/Auth/Specializations/SpecializationRequirement.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 
-namespace CopilotChat.WebApi.Auth;
+namespace CopilotChat.WebApi.Auth.Specializations;
 
 /// <summary>
 /// Used to require the specialization to be owned by the authenticated user.
